Create context package and move context-related class there
