+++ src/main/kotlin/api/handler/SocialRefresh.kt
@@ -6,7 +6,7 @@ import dev.deadzone.core.data.AdminData
 import dev.deadzone.module.logInput
 import dev.deadzone.module.logOutput
 import dev.deadzone.module.pioFraming
-import dev.deadzone.ServerContext
+import dev.deadzone.context.ServerContext
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.request.*
 import io.ktor.server.response.*
