+++ src/main/kotlin/module/Socket.kt
@@ -1,7 +1,7 @@
 package dev.deadzone.module
 
 import dev.deadzone.socket.Server
-import dev.deadzone.ServerContext
+import dev.deadzone.context.ServerContext
 
 fun configureSocket(context: ServerContext) {
     val server = Server(context = context)
