+++ src/main/kotlin/socket/Server.kt
@@ -1,6 +1,6 @@
 package dev.deadzone.socket
 
-import dev.deadzone.ServerContext
+import dev.deadzone.context.ServerContext
 import dev.deadzone.module.Logger
 import dev.deadzone.socket.handler.*
 import dev.deadzone.socket.tasks.TimeUpdate
