+++ src/main/kotlin/socket/tasks/TimeUpdate.kt
@@ -1,7 +1,7 @@
 package dev.deadzone.socket.tasks
 
 import dev.deadzone.socket.Connection
-import dev.deadzone.ServerContext
+import dev.deadzone.context.ServerContext
 import dev.deadzone.socket.utils.ServerPushTask
 import io.ktor.util.date.getTimeMillis
 import kotlinx.coroutines.delay
